requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-cors-headers==4.1.0
django-rest-authtoken==2.1.4
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==21.2.0
idna==3.4
importlib-resources==6.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
vcversioner==2.16.0.0
zipp==3.16.2
zope.interface==6.0
